+++ MQTTClient/MQTTClient/MQTTSessionLegacy.m
@@ -337,7 +337,7 @@
     return self.transport.host;
 }
 
-- (UInt16)port
+- (UInt32)port
 {
     return self.transport.port;
 }
